Seperated `smart` & `dumb` components
Major components refactoring
