+++ client/src/app/components/NarrowTable.js
@@ -16,37 +16,23 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import { useEffect, useState } from 'react';
 import Table from 'react-bootstrap/Table';
 import Spinner from 'react-bootstrap/Spinner';
 import styled, { createGlobalStyle } from 'styled-components';
-import getProducts from '../services/productService';
-import ExpandedProduct from './ExpandedProduct';
+import ProductEntry from './ProductEntry';
 import { faCaretDown } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 const TableContainer = createGlobalStyle`
   div.table-responsive {
     display: flex !important;
-    width: 100%;
-    padding: 0 0 0 10px !important;
-    margin: 0;
-    overflow: auto !important;
-
-    table {
-      width: 100vw !important;
-    }
-
-    #ItemCode {
-      text-align: center;
-    }
+    padding-left: 1rem !important;
   }
 `;
 
 const TableHeader = styled.th`
   color: #c3c3c3;
   border: none !important;
-  width: max-content !important;
 
   :hover {
     color: #1a8ac4;
@@ -55,13 +41,13 @@ const TableHeader = styled.th`
 `;
 
 const ProductTable = styled(Table)`
-  display: inline-table;
   height: 100%;
   width: 100%;
   min-width: 100% !important;
   min-height: 100%;
-  margin: 0 0 auto 0;
-  overflow: auto !important;
+  margin: 0 3.5rem 2rem 0;
+  overflow-x: scroll !important;
+  overflow-y: hidden;
   background-color: transparent;
   border-collapse: separate;
   border-spacing: 0 1rem;
@@ -73,7 +59,7 @@ const ProductTable = styled(Table)`
 `;
 
 // Loading Spinner
-const Loader = styled(Spinner)`
+export const Loader = styled(Spinner)`
   margin: 10px auto;
   width: 3rem;
   height: 3rem;
@@ -83,76 +69,42 @@ const Loader = styled(Spinner)`
 `;
 
 // Shows on no products registered
-const NullItems = styled.p`
+export const NullItems = styled.p`
   color: #c3c3c3;
   margin-bottom: 1rem;
   position: absolute;
-  left: 48%;
+  left: 36%;
   top: 16rem;
 `;
 
-export default function CatalogTable() {
-  const [products, setProducts] = useState(null);
-
-  useEffect(() => {
-    if (!products) {
-      fetchProducts().catch((err) => {
-        console.error(err);
-      });
-    }
-  });
-
-  const fetchProducts = async () => {
-    let res = await getProducts.getAll();
-    setProducts(res);
-  };
-
-  const verifyProducts = () => {
-    return (
-      <>
-        {products && products.length >= 1 ? (
-          products.map((products) => ExpandedProduct(products))
-        ) : (
-          <Loader variant="primary" animation="border" role="status" />
-        )}
-      </>
-    );
-  };
-
+const NarrowTable = (props) => {
   return (
     <>
       <TableContainer />
       <ProductTable hover responsive>
         <thead>
           <tr>
-            <TableHeader id="ItemCode">
+            <TableHeader colSpan="1">
               # Code <FontAwesomeIcon icon={faCaretDown} />
             </TableHeader>
-            <TableHeader id="ProductName">
+            <TableHeader>
               Product <FontAwesomeIcon icon={faCaretDown} />
             </TableHeader>
-            <TableHeader id="ItemType">
+            <TableHeader>
               Type <FontAwesomeIcon icon={faCaretDown} />
             </TableHeader>
-            <TableHeader id="Quantity">
+            <TableHeader>
               Quantity <FontAwesomeIcon icon={faCaretDown} />
-            </TableHeader>
-            <TableHeader id="CreatedAt">
-              Created <FontAwesomeIcon icon={faCaretDown} />
-            </TableHeader>
-            <TableHeader id="UpdatedAt">
-              Updated <FontAwesomeIcon icon={faCaretDown} />
             </TableHeader>
           </tr>
         </thead>
         <tbody>
-          {products && products.length === 0 ? (
-            <NullItems> No products registered...</NullItems>
-          ) : (
-            verifyProducts()
-          )}
+          <ProductEntry />
+          {props.data}
         </tbody>
       </ProductTable>
     </>
   );
-}
+};
+
+export default NarrowTable;
